Rearranged some of the layout files
Now the default layout file will be the tablet mode one (tested with emulator, must test with tablet before confirming) and the xxxhdpi is for high density device like the pixel so that we can run at the same time.
This change was made primarily it such that the main focus file (layout/activity_main) correspond to what we will be using in the leaderboard runs. Hence its there and we have an "additional" file for xxxhdpi. This is such that we are actually able to work at the same time when only 1 person has the tablet

Oh also extracted the strings because of it as well... Don't have to Ctrl+F and replace when dealing with 2 layouts

Signed-off-by: Kenneth Soh <sohy0037@e.ntu.edu.sg>