Adapted the layout to support phone as well (#1)
* Added an additional layout adjustment for mobile phones

The app will switch layout accordingly as and when the layout needs to change. So for tablets (which is our main point), it will use the layout file in layout-large instead. That will be our main layout file

Signed-off-by: Kenneth Soh <sohy0037@e.ntu.edu.sg>

* Made the dialog message box scale according to device dpi

This is to improve compatability among Android devices

Signed-off-by: Kenneth Soh <sohy0037@e.ntu.edu.sg>

* Added XML based icons for higher quality icons

Signed-off-by: Kenneth Soh <sohy0037@e.ntu.edu.sg>

* Switched over to using vector drawables officially

As the minSDK is 24 according to the gradle file, this means that the app actually natively supports vector drawables directly (support from API 21) and as such we do not need to maintain as many images as we are doing now. Should also improve build times as there are lesser files to compress and compact and stuff

Signed-off-by: Kenneth Soh <sohy0037@e.ntu.edu.sg>

* Rearranged some of the layout files

Now the default layout file will be the tablet mode one (tested with emulator, must test with tablet before confirming) and the xxxhdpi is for high density device like the pixel so that we can run at the same time.
This change was made primarily it such that the main focus file (layout/activity_main) correspond to what we will be using in the leaderboard runs. Hence its there and we have an "additional" file for xxxhdpi. This is such that we are actually able to work at the same time when only 1 person has the tablet

Oh also extracted the strings because of it as well... Don't have to Ctrl+F and replace when dealing with 2 layouts

Signed-off-by: Kenneth Soh <sohy0037@e.ntu.edu.sg>